Login : card, form, responsive
installation of card shadcn ui. integration of "Login card", 90% responsive. integration of card logo and image.

Need a code review to see if everything's ok so far.

TO DO (for tomorrow or for those who are free):
for responsive, we still need to define the minimum size of inputs, buttons, spans and images to prevent them from being too small before the breakpoint on lg to md and for sm to smaller.